5/4-5/5 changes
I have made the frame for the program. All i have to do is make it look
nicer with labels and then do the math for the program to work. Depends
how much space is left i might add a text box at the bottom that this
prints out to. All i need to figure out is gridlayout